+++ pcs_test/tier0/lib/cib/test_resource_common.py
@@ -1138,8 +1138,9 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
             self.validate(element, True),
             [
                 fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    report_codes.CANNOT_MOVE_RESOURCE_MASTER_RESOURCE_NOT_PROMOTABLE,
                     resource_id="R-clone",
+                    promotable_id="",
                 ),
             ],
         )
@@ -1148,12 +1149,7 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
         element = self._fixture_clone(False)
         assert_report_item_list_equal(
             self.validate(element, False),
-            [
-                fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
-                    resource_id="R-clone",
-                ),
-            ],
+            [],
         )
 
     def test_master_false_master(self):
@@ -1215,8 +1211,9 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
             self.validate(element.find("./primitive"), True),
             [
                 fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    report_codes.CANNOT_MOVE_RESOURCE_CLONE_INNER,
                     resource_id="R",
+                    clone_id="R-clone",
                 ),
             ],
         )
@@ -1227,8 +1224,9 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
             self.validate(element.find("./primitive"), False),
             [
                 fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    report_codes.CANNOT_MOVE_RESOURCE_CLONE_INNER,
                     resource_id="R",
+                    clone_id="R-clone",
                 ),
             ],
         )
@@ -1239,8 +1237,9 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
             self.validate(element.find("./group"), True),
             [
                 fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    report_codes.CANNOT_MOVE_RESOURCE_CLONE_INNER,
                     resource_id="G",
+                    clone_id="G-clone",
                 ),
             ],
         )
@@ -1251,8 +1250,9 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
             self.validate(element.find("./group"), False),
             [
                 fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    report_codes.CANNOT_MOVE_RESOURCE_CLONE_INNER,
                     resource_id="G",
+                    clone_id="G-clone",
                 ),
             ],
         )
@@ -1263,8 +1263,9 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
             self.validate(element.find("./group/primitive"), True),
             [
                 fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    report_codes.CANNOT_MOVE_RESOURCE_CLONE_INNER,
                     resource_id="R",
+                    clone_id="G-clone",
                 ),
             ],
         )
@@ -1275,8 +1276,9 @@ class ValidateMove(ValidateMoveBanClearMixin, TestCase):
             self.validate(element.find("./group/primitive"), False),
             [
                 fixture.error(
-                    report_codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    report_codes.CANNOT_MOVE_RESOURCE_CLONE_INNER,
                     resource_id="R",
+                    clone_id="G-clone",
                 ),
             ],
         )