+++ pcs_test/tier0/lib/commands/resource/test_resource_move_autoclean.py
@@ -672,10 +672,11 @@ class MoveAutocleanValidations(MoveAutocleanCommonSetup):
             ]
         )
 
-    def test_resource_wrong_type_clone(self):
-        resource_id = "clone_rsc"
+    def test_resource_wrong_type_clone_inner(self):
+        clone_id = "clone_rsc"
+        resource_id = "clone_rsc-primitive"
         self.config.runner.cib.load(
-            resources=_resources_tag(_rsc_clone_fixture(resource_id))
+            resources=_resources_tag(_rsc_clone_fixture(clone_id))
         )
         self.env_assist.assert_raise_library_error(
             lambda: move_autoclean(self.env_assist.get_env(), resource_id),
@@ -683,8 +684,9 @@ class MoveAutocleanValidations(MoveAutocleanCommonSetup):
         self.env_assist.assert_reports(
             [
                 fixture.error(
-                    reports.codes.CANNOT_MOVE_RESOURCE_CLONE,
+                    reports.codes.CANNOT_MOVE_RESOURCE_CLONE_INNER,
                     resource_id=resource_id,
+                    clone_id=clone_id,
                 )
             ]
         )
@@ -1549,11 +1551,14 @@ class MoveAutocleanFailures(MoveAutocleanCommonSetup):
             ]
         )
 
-    def assert_moving_bundle_with_more_instances(self, stderr, report_code):
+    def assert_moving_clone_bundle_with_more_instances(
+        self, stderr, report_code
+    ):
         # This tests that error messages from crm_resources are translated to
-        # report messages correctly. For the purpose of the test, no bundle is
-        # needed. The only thing that matters is the error message from
-        # crm_resource. So we don't bother creating a CIB with a bundle in it.
+        # report messages correctly. For the purpose of the test, no clone or
+        # bundle is needed. The only thing that matters is the error message
+        # from crm_resource. So we don't bother creating a CIB with a clone or
+        # bundle in it.
         self.tmp_file_mock_obj.set_calls(
             self.get_tmp_files_mocks(
                 _simulation_transition_fixture(
@@ -1585,14 +1590,14 @@ class MoveAutocleanFailures(MoveAutocleanCommonSetup):
         )
         self.env_assist.assert_reports(self.get_reports(stage=1))
 
-    def test_moving_bundle_with_more_instances_1(self):
-        self.assert_moving_bundle_with_more_instances(
+    def test_moving_clone_bundle_with_more_instances_1(self):
+        self.assert_moving_clone_bundle_with_more_instances(
             "Error performing operation: Multiple items match request",
             reports.codes.CANNOT_MOVE_RESOURCE_MULTIPLE_INSTANCES,
         )
 
-    def test_moving_bundle_with_more_instances_2(self):
-        self.assert_moving_bundle_with_more_instances(
+    def test_moving_clone_bundle_with_more_instances_2(self):
+        self.assert_moving_clone_bundle_with_more_instances(
             "Resource 'A' not moved: active in 2 locations.",
             reports.codes.CANNOT_MOVE_RESOURCE_MULTIPLE_INSTANCES_NO_NODE_SPECIFIED,
         )